Refactor to remove view holders
