+++ app/build.gradle
@@ -69,7 +69,7 @@ dependencies {
     implementation 'com.yesferal.hornsapp:hada:0.2.0'
     implementation 'androidx.viewpager2:viewpager2:1.0.0'
     implementation 'com.squareup.picasso:picasso:2.71828'
-    implementation 'com.google.android.material:material:1.3.0-alpha03'
+    implementation 'com.google.android.material:material:1.3.0-alpha04'
     implementation 'com.google.firebase:firebase-analytics-ktx:18.0.0'
     implementation 'com.google.android.gms:play-services-ads:19.5.0'
     implementation 'com.google.firebase:firebase-crashlytics:17.3.0'
