--- app/src/main/java/com/yesferal/hornsapp/app/presentation/ui/concert/detail/adapter/BandViewHolder.kt
@@ -1,25 +1,0 @@
-package com.yesferal.hornsapp.app.presentation.ui.concert.detail.adapter
-
-import android.view.View
-import com.yesferal.hornsapp.app.presentation.common.custom.load
-import com.yesferal.hornsapp.app.presentation.common.custom.setAllCornersRounded
-import com.yesferal.hornsapp.app.presentation.common.multitype.BaseViewHolder
-import com.yesferal.hornsapp.app.presentation.common.multitype.ViewHolderBinding
-import com.yesferal.hornsapp.app.presentation.ui.concert.detail.BandViewData
-import kotlinx.android.synthetic.main.item_band.view.*
-
-class BandViewHolder (
-    itemView: View,
-    private val listener: ViewHolderBinding.Listener
-) : BaseViewHolder<BandViewData>(itemView) {
-
-    override fun bind(viewData: BandViewData) {
-        itemView.itemTextView.text = viewData.name
-        itemView.itemImageView.setAllCornersRounded()
-        itemView.itemImageView.load(viewData.membersImage)
-
-        itemView.itemImageView.setOnClickListener {
-            (listener as BandViewData.Listener).onClick(viewData)
-        }
-    }
-}
