--- app/src/main/java/com/yesferal/hornsapp/app/presentation/ui/concert/upcoming/adapter/ErrorViewHolder.kt
@@ -1,18 +1,0 @@
-package com.yesferal.hornsapp.app.presentation.ui.concert.upcoming.adapter
-
-import android.view.View
-import com.yesferal.hornsapp.app.presentation.common.multitype.BaseViewHolder
-import com.yesferal.hornsapp.app.presentation.common.custom.setUpWith
-import com.yesferal.hornsapp.app.presentation.ui.concert.upcoming.ErrorViewData
-import kotlinx.android.synthetic.main.custom_error.view.*
-
-class ErrorViewHolder(
-    itemView: View
-) : BaseViewHolder<ErrorViewData>(itemView) {
-
-    override fun bind(viewData: ErrorViewData) {
-        itemView.errorTextView.let {
-            it.setUpWith(it.context.getString(viewData.errorMessage))
-        }
-    }
-}
